Moving datastreams call to Shell
