+++ src/screens/Shell.tsx
@@ -1,9 +1,16 @@
+import { useEffect } from "react";
+import { useAuth } from "react-oidc-context";
 import {
   MenuItemType,
   PopupMenuItemType,
 } from "@us-gov-cdc/cdc-react/dist/src/@types";
 
 import dexLogo from "src/assets/dex_logo.svg";
+import { useSetRecoilState } from "recoil";
+import { dataRouteAtom, dataStreamIdAtom } from "src/state/searchParams";
+import { dataStreamsAtom } from "src/state/dataStreams";
+import getDataStreams, { DataStream } from "src/utils/api/dataStreams";
+import { getDataRoutes } from "src/utils/helperFunctions/dataStreams";
 
 import {
   ProfileHeader,
@@ -17,8 +24,31 @@ import { Icons } from "@us-gov-cdc/cdc-react-icons";
 
 function Shell() {
   const navigate = useNavigate();
+  const auth = useAuth();
 
   const logo = <ProfileHeaderLogo image={dexLogo} classNames={["dex-logo"]} />;
+
+  const setDataStreams = useSetRecoilState(dataStreamsAtom);
+  const setDataStream = useSetRecoilState(dataStreamIdAtom);
+  const setDataRoute = useSetRecoilState(dataRouteAtom);
+  useEffect(() => {
+    const fetchCall = async () => {
+      const res = await getDataStreams(auth.user?.access_token || "");
+
+      try {
+        const data = await res.json();
+        const streams = data?.dataStreams as DataStream[];
+        const dataStreamId = streams[0].dataStreamId;
+        setDataStreams(streams);
+        setDataStream(dataStreamId);
+        const route = getDataRoutes(streams, dataStreamId)[0];
+        setDataRoute(route);
+      } catch (error) {
+        console.error("Failed to parse JSON:", error);
+      }
+    };
+    fetchCall();
+  }, [auth]);
 
   const menuItems: MenuItemType[] = [
     {
@@ -118,7 +148,7 @@ function Shell() {
                   componentType: "a",
                   icon: <Icons.SquareHalfArrowRight />,
                   text: "Logout",
-                  href: "/logout",
+                  onClick: "/logout",
                 },
               ],
             },
